+++ server/mdm/acme/internal/tests/integration_test.go
@@ -1,7 +1,6 @@
 package tests
 
 import (
-	"crypto/ecdsa"
 	"fmt"
 	"net/http"
 	"testing"
@@ -800,25 +799,6 @@ func testCreateOrder(t *testing.T, s *integrationTestSuite) {
 		require.NotEmpty(t, resp.Header.Get("Replay-Nonce"))
 		require.Empty(t, resp.Header.Get("Location"))
 	})
-}
-
-// createOrderForGet is a convenience helper that creates an account and order for an enrollment,
-// returning the private key, account URL, order response, and a fresh nonce for subsequent requests.
-func (s *integrationTestSuite) createOrderForGet(t *testing.T, enroll *types.Enrollment) (*ecdsa.PrivateKey, string, *types.OrderResponse, string) {
-	t.Helper()
-	privateKey, accountURL, nonce := s.createAccountForOrder(t, enroll)
-
-	payload := map[string]any{
-		"identifiers": []map[string]string{
-			{"type": "permanent-identifier", "value": enroll.HostIdentifier},
-		},
-	}
-	jwsBody := buildJWS(t, privateKey, nonce, accountURL, s.newOrderURL(enroll.PathIdentifier), payload)
-	orderResp, _, resp := s.createOrder(t, enroll.PathIdentifier, jwsBody)
-	require.Equal(t, http.StatusCreated, resp.StatusCode)
-	nextNonce := resp.Header.Get("Replay-Nonce")
-	require.NotEmpty(t, nextNonce)
-	return privateKey, accountURL, orderResp, nextNonce
 }
 
 func testGetOrder(t *testing.T, s *integrationTestSuite) {
